// src/api/generated/schemas/travelTypeCode.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */

export type TravelTypeCode = typeof TravelTypeCode[keyof typeof TravelTypeCode];


export const TravelTypeCode = {
  adventurer: 'adventurer',
  explorer: 'explorer',
  relaxer: 'relaxer',
  foodie: 'foodie',
} as const;